LPS-128599 Move the implementations as inner classes of Log4JLoggerTestUtil
